requirements.txt
# Fraudasaurus.ai - Jack Henry DevCon 2026 Hack-A-Thon
# Core data
pandas>=2.0
numpy>=1.24
pyarrow>=14.0
openpyxl>=3.1

# Machine learning
scikit-learn>=1.3

# Visualization
matplotlib>=3.7
seaborn>=0.13
plotly>=5.18

# Graph analysis
networkx>=3.1

# Notebooks
jupyter>=1.0
ipywidgets>=8.0

# Report generation
jinja2>=3.1
tabulate>=0.9

# BigQuery
google-cloud-bigquery>=3.0
google-cloud-bigquery-storage>=2.0
db-dtypes>=1.0

# Utilities
tqdm>=4.66
